refactor(frontend): split API service layer
