+++ frontend/package.json
@@ -15,25 +15,28 @@
     "format:fix": "prettier . --write",
     "lint:styles": "stylelint \"**/*.{css,scss}\" --cache",
     "lint:styles:fix": "stylelint \"**/*.{css,scss}\" --fix --cache",
-    "check": "npm run lint && npm run format && npm run lint:styles && npm run build",
+    "test": "vitest",
+    "test:run": "vitest run",
+    "check": "npm run lint && npm run format && npm run lint:styles && npm run test:run && npm run build",
     "prepare": "node ./prepare-husky.mjs",
     "preview": "vite preview"
   },
   "dependencies": {
     "@gravity-ui/navigation": "^2.33.0",
     "@gravity-ui/uikit": "^6.43.0",
-    "@simplewebauthn/browser": "^13.1.2",
     "axios": "^1.15.0",
     "bootstrap": "^5.3.3",
     "prop-types": "^15.8.1",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.9",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^7.12.0",
-    "react-toastify": "^11.0.3"
+    "react-router-dom": "^7.12.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.19.0",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.2",
+    "@testing-library/user-event": "^14.6.1",
     "@types/react": "^19.0.8",
     "@types/react-dom": "^19.0.3",
     "@vitejs/plugin-react-swc": "^3.5.0",
@@ -46,10 +49,13 @@
     "eslint-plugin-simple-import-sort": "^13.0.0",
     "globals": "^15.14.0",
     "husky": "^9.1.7",
+    "jsdom": "^29.0.2",
     "lint-staged": "^16.4.0",
+    "msw": "^2.13.6",
     "prettier": "^3.5.0",
     "stylelint": "^16.14.1",
     "stylelint-config-standard": "^37.0.0",
-    "vite": "^6.4.2"
+    "vite": "^6.4.2",
+    "vitest": "^4.1.5"
   }
 }
