+++ frontend/src/services/api.js
@@ -1,0 +1,1 @@
+export * from "./api/index";
