+++ frontend/vite.config.js
@@ -4,6 +4,12 @@ import { defineConfig } from "vite";
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  test: {
+    environment: "jsdom",
+    setupFiles: ["./src/test/setup.js"],
+    restoreMocks: true,
+    clearMocks: true,
+  },
   build: {
     rollupOptions: {
       output: {
